backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apiflask>=3.0.2",
    "argon2-cffi>=25.1.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
]


[tool.mypy]
plugins = ["pydantic.mypy"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.15.0",
]

# [tool.pydantic-mypy]
# init_forbid_extra = true
# init_typed = true
# warn_required_dynamic_aliases = true
